Small clean-ups.
git-svn-id: https://svn.dealii.org/trunk@8111 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/fe/fe_dgp.cc
@@ -193,7 +193,7 @@ template <int dim>
 std::vector<unsigned int>
 FE_DGP<dim>::get_dpo_vector(unsigned int deg)
 {
-  std::vector<unsigned int> dpo(dim+1, static_cast<unsigned int>(0));
+  std::vector<unsigned int> dpo(dim+1, 0U);
   dpo[dim] = ++deg;
   for (unsigned int i=1;i<dim;++i)
     {
